+++ app/Http/Controllers/BookController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\UpdateBookRequest;
 use App\Models\Author;
 use App\Models\Book;
 use App\Models\Category;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
@@ -68,10 +69,12 @@ class BookController extends Controller
 
     /**
      * @param Book $book
-     * @return Application|Factory|View|\Illuminate\Foundation\Application
+     * @return \Illuminate\Foundation\Application|View|Factory|Application
+     * @throws AuthorizationException
      */
     public function edit(Book $book): \Illuminate\Foundation\Application|View|Factory|Application
     {
+        $this->authorize('update', $book);
         $authors = Author::all();
         $categories = Category::all();
 
@@ -87,9 +90,11 @@ class BookController extends Controller
      * @param UpdateBookRequest $request
      * @param Book $book
      * @return RedirectResponse
+     * @throws AuthorizationException
      */
     public function update(UpdateBookRequest $request, Book $book): RedirectResponse
     {
+        $this->authorize('update', $book);
         $data = $request->all();
         $file = $request->file('image');
 
@@ -108,9 +113,11 @@ class BookController extends Controller
     /**
      * @param Book $book
      * @return RedirectResponse
+     * @throws AuthorizationException
      */
     public function destroy(Book $book): RedirectResponse
     {
+        $this->authorize('delete', $book);
         $book->delete();
 
         return redirect()
